refactor: move stop-string detokenizer to atom/entrypoints
Stop strings are matched on detokenized text, and text only exists where a
tokenizer does -- the frontend. The module sat in atom/model_engine, the
package the GPU workers import.

Move it to atom/entrypoints/detokenizer.py and take the wrapper body with it
as a free function `wrap_for_stop_strings`. `InputOutputProcessor` keeps the
single call site every request path passes through and now delegates through
a lazy import, so `atom.model_engine` still pulls nothing from the serving
package at module load.

`InputOutputProcessor` itself cannot follow: `LLMEngine.__init__` builds it,
and `ModelRunner` imports it for the per-req-cache startup check in the GPU
worker.

No behaviour change beyond `abort_request` now being bound when the callback
is wrapped rather than read when the abort fires; `LLMEngine.__init__` wires
it before any request is preprocessed.
